Added github actions
This will compile, format, vet and test the code when it is pushed directly or pushed to a PR.
The build step will check for compile errors.
The test step will ensure that all tests pass.
The gofmt step will ensure that the code is formatted.
The vet step will apply some heuristics to spot common bugs and style errors.